backend/requirements.txt
cffi==1.16.0
databricks-sql-connector
Faker==25.0.1
fastapi==0.111.0
fpdf==1.7.2
httpx==0.27.0
protobuf
pydantic
pydub==0.25.1
pymongo[srv]
pytest==7.0.1
python-dotenv==1.0.1
pyttsx3
qrcode==7.4.2
Requests==2.31.0
resend==0.8.0
sounddevice==0.4.6
streamlit==1.24.0
streamlit_pdf_viewer==0.0.10
typing_extensions==4.11.0
uvicorn==0.29.0
websockets
hume
hume[microphone]
SpeechRecognition[whisper-local]